+++ internal/models/template/noise.go
@@ -1,6 +1,8 @@
 package template
 
+// Noise describes a single noise layer applied to borders/water.
+// `amp` may be integer or fractional across the example templates.
 type Noise struct {
-	Amplitude int `json:"amp"`
-	Frequency int `json:"freq"`
+	Amplitude float64 `json:"amp"`
+	Frequency int     `json:"freq"`
 }
